+++ app/models/Book.scala
@@ -3,7 +3,7 @@ package models
 import play.api.http.Writeable
 import play.api.libs.json.{JsValue, Json}
 
-case class Book(author: String, title: String, year: Int, code: String, id: Option[Int] = None)
+case class Book(author: String, title: String, year: Int, code: String, id: Option[Long] = None)
 
 object Book {
   implicit val format = Json.format[Book]
