+++ app/models/Visitor.scala
@@ -3,7 +3,7 @@ package models
 import play.api.http.Writeable
 import play.api.libs.json._
 
-case class Visitor(firstName: String, lastName: String, middleName: Option[String], extraName: Option[String], id: Option[Int] = None)
+case class Visitor(firstName: String, lastName: String, middleName: Option[String], extraName: Option[String], id: Option[Long] = None)
 
 object Visitor {
   implicit val format = Json.format[Visitor]
